+++ src/main/java/com/library/Service/FavoriteUserService.java
@@ -1,11 +1,15 @@
 package com.library.Service;
 
+import com.library.Entity.Book;
 import com.library.Entity.FavoriteUser;
 import com.library.Repository.*;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
 
 @Service
@@ -20,20 +24,53 @@ public class FavoriteUserService {
         this.favorite_user = favoriteUser;
     }
 
-    public List<FavoriteUser> ListOfFavoriteBooks() {
-        return favorite_user.findAll();
+    public ResponseEntity<List<FavoriteUser>> ListOfFavoriteBooks() {
+        try{
+            List<FavoriteUser>favoriteUsers = favorite_user.findAll();
+            return new ResponseEntity<>(favoriteUsers, HttpStatus.OK);
+        }catch (Exception e)
+        {
+            return new ResponseEntity<>(Collections.emptyList(),HttpStatus.NO_CONTENT);
+        }
     }
 
-    public FavoriteUser AddBooksToFavoriteList(FavoriteUser Favorite_user) {
-        return favorite_user.save(Favorite_user);
+    public ResponseEntity<HttpStatus> AddBooksToFavoriteList(FavoriteUser Favoriteuser) {
+        try{
+            if(Favoriteuser == null||Favoriteuser.getUser()==null||Favoriteuser.getBook()==null)
+            {
+                return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+            }
+            favorite_user.save(Favoriteuser);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        }catch (Exception e)
+        {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
-    public List<FavoriteUser> ListOfFavoriteBooksByUsername(String Username) {
-        return favorite_user.findByUser_Username(Username);
+    public ResponseEntity<List<FavoriteUser>> ListOfFavoriteBooksByUsername(String Username) {
+        try{
+            List<FavoriteUser>favoriteUsers = favorite_user.findByUser_Username(Username);
+            return new ResponseEntity<>(favoriteUsers, HttpStatus.OK);
+        }catch (Exception e)
+        {
+            return new ResponseEntity<>(Collections.emptyList(),HttpStatus.NO_CONTENT);
+        }
     }
 
     @Transactional
-    public void Delete(int id) {
-        favorite_user.deleteById(id);
+    public ResponseEntity<HttpStatus> Delete(int id) {
+        try {
+            if(favorite_user.existsById(id)) {
+                favorite_user.deleteById(id);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }else {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            }
+
+        }catch(Exception e){
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
     }
 
 }
